+++ lib/flutter_recorder.dart
@@ -3,4 +3,7 @@ library flutter_recorder;
 
 export 'src/audio_data_container.dart';
 export 'src/enums.dart' show PCMFormat, RecorderChannels;
+export 'src/filters/autogain.dart';
+export 'src/filters/echo_cancellation.dart';
+export 'src/filters/filters.dart' show RecorderFilterType;
 export 'src/flutter_recorder.dart';