+++ lib/src/bindings/js_extension.dart
@@ -6,32 +6,32 @@ import 'package:web/web.dart' as web;
 // //////////////////////////
 // common
 // //////////////////////////
-@JS('Module._malloc')
+@JS('RecorderModule._malloc')
 external int wasmMalloc(int bytesCount);
 
-@JS('Module._free')
+@JS('RecorderModule._free')
 external void wasmFree(int ptrAddress);
 
-@JS('Module.getValue')
+@JS('RecorderModule.getValue')
 external int wasmGetI32Value(int ptrAddress, String type);
 
-@JS('Module.getValue')
+@JS('RecorderModule.getValue')
 external double wasmGetF32Value(int ptrAddress, String type);
 
-@JS('Module.UTF8ToString')
+@JS('RecorderModule.UTF8ToString')
 external String wasmUtf8ToString(int ptrAddress);
 
-@JS('Module.setValue')
+@JS('RecorderModule.setValue')
 external void wasmSetValue(int ptrAddress, int value, String type);
 
-@JS('Module.cwrap')
+@JS('RecorderModule.cwrap')
 external JSFunction wasmCwrap(
   JSString fName,
   JSString returnType,
   JSArray<JSString> argTypes,
 );
 
-@JS('Module.ccall')
+@JS('RecorderModule.ccall')
 external JSFunction wasmCccall(
   JSString fName,
   JSString returnType,
@@ -39,37 +39,37 @@ external JSFunction wasmCccall(
   JSArray<JSAny> args,
 );
 
-@JS('Module._createWorkerInWasm')
+@JS('RecorderModule._createWorkerInWasm')
 external void wasmCreateWorkerInWasm();
 
-@JS('Module._setDartEventCallback')
+@JS('RecorderModule._setDartEventCallback')
 external void wasmSetDartEventCallback(int callbackPtr, int callbackPtr2);
 
-@JS('Module._sendToWorker')
+@JS('RecorderModule._sendToWorker')
 external void wasmSendToWorker(int message, int value);
 
-@JS('Module.wasmWorker')
+@JS('RecorderModule.wasmWorker')
 external web.Worker wasmWorker;
 
 // //////////////////////////
 // bindings
 // //////////////////////////
 
-@JS('Module._setSilenceDetection')
+@JS('RecorderModule._setSilenceDetection')
 external void wasmSetSilenceDetection(bool enable);
 
-@JS('Module._setSilenceThresholdDb')
+@JS('RecorderModule._setSilenceThresholdDb')
 external void wasmSetSilenceThresholdDb(double silenceThresholdDb);
 
-@JS('Module._setSilenceDuration')
+@JS('RecorderModule._setSilenceDuration')
 external void wasmSetSilenceDuration(double silenceDuration);
 
-@JS('Module._setSecondsOfAudioToWriteBefore')
+@JS('RecorderModule._setSecondsOfAudioToWriteBefore')
 external void wasmSetSecondsOfAudioToWriteBefore(
   double secondsOfAudioToWriteBefore,
 );
 
-@JS('Module._listCaptureDevices')
+@JS('RecorderModule._listCaptureDevices')
 external void wasmListCaptureDevices(
   int namesPtr,
   int deviceIdPtr,
@@ -77,7 +77,7 @@ external void wasmListCaptureDevices(
   int nDevicePtr,
 );
 
-@JS('Module._freeListCaptureDevices')
+@JS('RecorderModule._freeListCaptureDevices')
 external void wasmFreeListCaptureDevices(
   int namesPtr,
   int deviceIdPtr,
@@ -85,53 +85,79 @@ external void wasmFreeListCaptureDevices(
   int nDevicePtr,
 );
 
-@JS('Module._init')
+@JS('RecorderModule._init')
 external int wasmInit(int deviceID, int format, int sampleRate, int channels);
 
-@JS('Module._deinit')
+@JS('RecorderModule._deinit')
 external void wasmDeinit();
 
-@JS('Module._isInited')
+@JS('RecorderModule._isInited')
 external int wasmIsDeviceInitialized();
 
-@JS('Module._isDeviceStarted')
+@JS('RecorderModule._isDeviceStarted')
 external int wasmIsDeviceStarted();
 
-@JS('Module._start')
+@JS('RecorderModule._start')
 external int wasmStart();
 
-@JS('Module._stop')
+@JS('RecorderModule._stop')
 external void wasmStop();
 
-@JS('Module._startStreamingData')
+@JS('RecorderModule._startStreamingData')
 external void wasmStartStreamingData();
 
-@JS('Module._stopStreamingData')
+@JS('RecorderModule._stopStreamingData')
 external void wasmStopStreamingData();
 
-@JS('Module._startRecording')
+@JS('RecorderModule._startRecording')
 external int wasmStartRecording(int pathPtr);
 
-@JS('Module._setPauseRecording')
+@JS('RecorderModule._setPauseRecording')
 external void wasmSetPauseRecording(bool pause);
 
-@JS('Module._stopRecording')
+@JS('RecorderModule._stopRecording')
 external void wasmStopRecording();
 
-@JS('Module._setFftSmoothing')
+@JS('RecorderModule._setFftSmoothing')
 external void wasmSetFftSmoothing(double smooth);
 
-@JS('Module._getFft')
+@JS('RecorderModule._getFft')
 external void wasmGetFft(int samplesPtr);
 
-@JS('Module._getWave')
+@JS('RecorderModule._getWave')
 external void wasmGetWave(int samplesPtr);
 
-@JS('Module._getTexture')
+@JS('RecorderModule._getTexture')
 external void wasmGetTexture(int samplesPtr);
 
-@JS('Module._getTexture2D')
+@JS('RecorderModule._getTexture2D')
 external void wasmGetTexture2D(int samplesPtr);
 
-@JS('Module._getVolumeDb')
+@JS('RecorderModule._getVolumeDb')
 external void wasmGetVolumeDb(int volumeDbPtr);
+
+@JS('RecorderModule._isFilterActive')
+external int wasmIsFilterActive(int filterTypeId);
+
+@JS('RecorderModule._addFilter')
+external int wasmAddFilter(int filterTypeId);
+
+@JS('RecorderModule._removeFilter')
+external int wasmRemoveFilter(int filterTypeId);
+
+@JS('RecorderModule._getFilterParamNames')
+external int wasmGetFilterParamNames(
+  int filterTypeId,
+  int namesPtr,
+  int paramsCountPtr,
+);
+
+@JS('RecorderModule._setFilterParams')
+external void wasmSetFilterParamValue(
+  int filterTypeId,
+  int attributeId,
+  double value,
+);
+
+@JS('RecorderModule._getFilterParams')
+external double wasmGetFilterParamValue(int filterTypeId, int attributeId);