+++ lib/src/exceptions/exceptions.dart
@@ -59,15 +59,21 @@ abstract class RecorderCppException extends RecorderException {
       case CaptureErrors.captureInitFailed:
         return const RecorderInitializeFailedException();
       case CaptureErrors.captureNotInited:
-        return const RecorderCaptureNotInitializedException();
+        return const RecorderNotInitializedException();
       case CaptureErrors.failedToInitializeRecording:
         return const RecorderFailedToInitializeRecordingException();
       case CaptureErrors.failedToStartDevice:
         throw const RecorderFailedToStartDeviceException();
-      case CaptureErrors.failedToWriteWav:
-        return const RecorderFailedToWriteWavException();
       case CaptureErrors.invalidArgs:
         return const RecorderInvalidArgumentsException();
+      case CaptureErrors.failedToWriteWav:
+        return const RecorderFailedToWriteWavException();
+      case CaptureErrors.filterNotFound:
+        return const RecorderFilterNotFoundException();
+      case CaptureErrors.filterAlreadyAdded:
+        return const RecorderFilterAlreadyAddedException();
+      case CaptureErrors.filterParameterGetError:
+        return const RecorderFilterParameterGetException();
     }
   }
 }